users page

+++ actions/deleteCategory.php
@@ -1,11 +1,7 @@
 <?php
-    session_start();
-    require_once("../utils/utils.php");
-    if(is_logged_in() && is_admin()) {
-        $id = $_GET['id'];
+    if(is_logged_in() && is_admin() && isset($_POST['action']) && $_POST['action'] == 'delete') {
+        $id = $_POST['id'];
         $sql = "delete from categories where id = ?";
-        require_once("../utils/server.php");
         $result = $connection->execute_query($sql, array($id));
-        redirect("categories.php", $url);
     }
 ?>
